CMakeLists

+++ src/hypua2jamo-c/p2jd2-table.h
@@ -1,4 +1,4 @@
-typedef unsigned int codepoint_t;
+typedef unsigned short codepoint_t;
 static const codepoint_t pua2jamo_E0BC[] = { 0x0003, 0x115f, 0x1161, 0x11ae };
 static const codepoint_t pua2jamo_E0BD[] = { 0x0004, 0x115f, 0x1161, 0x11ae, 0x11ae };
 static const codepoint_t pua2jamo_E0BE[] = { 0x0003, 0x115f, 0x1161, 0x11af };
